feat: map expanded gateway network data
