+++ src/MercadoPago/Resources/Payment/Expanded.php
@@ -30,41 +30,3 @@ class Expanded
         return $this->map;
     }
 }
-
-/**
- * Represents the gateway section of expanded payment data.
- *
- * Contains reference information from the payment gateway/acquirer.
- */
-class ExpandedGateway
-{
-    /** Maps nested objects to their corresponding DTO classes. */
-    use Mapper;
-
-    /** @var ExpandedGatewayReference|null Reference identifiers returned by the gateway. */
-    public ?object $reference;
-
-    private $map = [
-        "reference" => "MercadoPago\Resources\Payment\ExpandedGatewayReference",
-    ];
-
-    /**
-     * Method responsible for getting map of entities.
-     */
-    public function getMap(): array
-    {
-        return $this->map;
-    }
-}
-
-/**
- * Represents gateway reference identifiers from the card network.
- *
- * Used for recurring payment flows and acquirer reconciliation when
- * a network transaction ID is required for subsequent transactions.
- */
-class ExpandedGatewayReference
-{
-    /** Unique transaction identifier assigned by the card network (e.g. Visa, Mastercard). */
-    public ?string $network_transaction_id;
-}
