+++ tests/MercadoPago/Resources/Payment/NetworkDataTest.php
@@ -4,6 +4,7 @@ namespace MercadoPago\Tests\Resources\Payment;
 
 use MercadoPago\Resources\Payment\PointOfInteraction;
 use MercadoPago\Resources\Payment\TransactionData;
+use MercadoPago\Resources\Payment\ExpandedGatewayReference;
 use PHPUnit\Framework\TestCase;
 
 final class NetworkDataTest extends TestCase
@@ -15,5 +16,12 @@ final class NetworkDataTest extends TestCase
 
         $this->assertSame('MercadoPago\\Resources\\Payment\\NetworkData', $transactionData->getMap()['network_data']);
         $this->assertArrayNotHasKey('network_data', $pointOfInteraction->getMap());
+    }
+
+    public function testExpandedGatewayReferenceMapsNetworkData(): void
+    {
+        $reference = new ExpandedGatewayReference();
+
+        $this->assertSame('MercadoPago\\Resources\\Payment\\NetworkData', $reference->getMap()['network_data']);
     }
 }
